pyproject.toml
[project]
name = "kim-bot"
version = "0.1.0"
description = "rag llm"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alive-progress>=3.3.0",
    "chromadb>=1.4.1",
    "click>=8.3.1",
    "docling>=2.70.0",
    "ollama>=0.6.1",
    "pylibmagic>=0.5.0",
    "python-magic>=0.4.27",
    "transformers>=4.57.6",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
]

[project.scripts]
kim-bot = "kim-bot.main:cli"